feat: add Content Security Policy
Aim is to lock everything down then loosen as necessary.

Using [web.dev Content security policy][1] as primary guide, and
[Google's CSP Evaluator][2] as primary review tool.

CSP suggests two additions to this policy:

1. `script-src`: Consider using `strict-dynamic` in combination with
CSP nonces or hashes.
2. `require-trusted-types-for`: Consider requiring Trusted Types for
scripts to lock down DOM XSS injection sinks. You can do this by
adding `require-trusted-types-for 'script'` to your policy.

I'm skipping both for now. The first is not really necessary for this
site as it stands and the second is experimental (but worth keeping an
eye on in future, though).

[1]: https://web.dev/csp/
[2]: https://csp-evaluator.withgoogle.com/

feat: remove cache control on font